refact: update public interface librpa.h
interface:
- remove `set_dimension`, add `set_atomic_basis`
- add `nbasis` arguemtn to `set_wg_ekb_efermi`
- remove basis size and insert_index_only arguments from `set_ao_basis_aux`

src:
- change argument types of methods of `AtomicBasis`.
  Rules:
  * atom index: `atom_t`
  * basis size: `size_t`
  * basis index: `int`
- change `atom_t` to `int` from `size_t` in `atoms.h`

driver:
- add `read_basis_out` to read atomic basis information from `basis_out`
- change Cs binary reader to use bytes offsets in the header for quick
  filtering
- adapt Cs and meanfield reader to changes in the interface

+++ src/atoms.h
@@ -19,7 +19,7 @@ using std::pair;
 using std::size_t;
 
 //! type of atom indices
-typedef size_t atom_t;
+typedef int atom_t;
 
 //! atom-pair type. NOTE: may turn into a class?
 typedef pair<atom_t, atom_t> atpair_t;
